+++ lib/hexapdf/filter/flate_decode.rb
@@ -87,9 +87,9 @@ module HexaPDF
         end
 
         Fiber.new do
-          deflater = Zlib::Deflate.new(HexaPDF::GlobalConfiguration['filter.flate_compression'],
+          deflater = Zlib::Deflate.new(HexaPDF::GlobalConfiguration['filter.flate.compression'],
                                        Zlib::MAX_WBITS,
-                                       HexaPDF::GlobalConfiguration['filter.flate_memory'])
+                                       HexaPDF::GlobalConfiguration['filter.flate.memory'])
           while source.alive? && (data = source.resume)
             data = deflater.deflate(data)
             Fiber.yield(data)